cpu-on-demand: Add main introspection abilities
For now we're able to enumerate all of our CPU devices and determine which
driver they're using, and then transform the mode to the correct governor
name for the given driver (i.e. `intel_pstate` vs `cpufreq`).

For now we're not setting the new governor, that'll come in the next step.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>